MAINT: Moved all detrending-related code to the respective `Detrending` modules; restructured detrending for uniform line scans
