+++ SurfaceTopography/NonuniformLineScan.py
@@ -252,140 +252,6 @@ class StaticallyScaledNonuniformTopography(ScaledNonuniformTopography):
         return self._position_scale_factor
 
 
-class DetrendedNonuniformTopography(DecoratedNonuniformTopography):
-    """
-    Remove trends from a topography. This is achieved by fitting polynomials
-    to the topography data to extract trend lines. The resulting topography
-    is then detrended by substracting these trend lines.
-    """
-
-    _detrend_functions = {
-        'mean': lambda self: [self.parent_topography.mean()],
-        # same as 'mean', deprecate 'center' in the future
-        'center': lambda self: [self.parent_topography.mean()],
-        'median': lambda self: [self.parent_topography.median()],
-        'rms-tilt': lambda self: self.parent_topography.polyfit(1),
-        # same as 'rms-tilt', deprecate 'height' in the future
-        'height': lambda self: self.parent_topography.polyfit(1),
-        'mad-tilt': lambda self: self.parent_topography.mad_polyfit(1),
-        'slope': lambda self: [self.parent_topography.mean(), self.parent_topography.derivative(1).mean()],
-        'rms-curvature': lambda self: self.parent_topography.polyfit(2),
-        # same as 'rms-curvature', deprecate 'curvature' in the future
-        'curvature': lambda self: self.parent_topography.polyfit(2),
-        'mad-curvature': lambda self: self.parent_topography.mad_polyfit(2),
-    }
-
-    def __init__(self, topography, detrend_mode='height', coeffs=None, info={}):
-        """
-        Parameters
-        ----------
-        topography : SurfaceTopography
-            SurfaceTopography to be detrended.
-        detrend_mode : str
-            'mean': center the topography to its mean, no trend correction.
-            'median': center the topography to its median, no trend correction.
-            'rms-tilt': adjust slope such that rms height is minimized.
-            'mad-tilt': adjust slope such that mad height is minimized.
-            'slope': adjust slope such that rms slope is minimized.
-            'rms-curvature': adjust slope and curvature such that rms height is minimized.
-            (Default: 'height')
-        """
-        super().__init__(topography, info=info)
-        self._detrend_mode = detrend_mode
-        self._coeffs = coeffs
-        if self._coeffs is None:
-            self._detrend()
-
-    def _detrend(self):
-        try:
-            self._coeffs = self._detrend_functions[self._detrend_mode](self)
-        except KeyError:
-            raise ValueError("Unsupported detrend mode '{}' for line scans.".format(self._detrend_mode))
-
-    def __getstate__(self):
-        """ is called and the returned object is pickled as the contents for
-            the instance
-        """
-        state = super().__getstate__(), self._detrend_mode, self._coeffs
-        return state
-
-    def __setstate__(self, state):
-        """ Upon unpickling, it is called with the unpickled state
-        Keyword Arguments:
-        state -- result of __getstate__
-        """
-        superstate, self._detrend_mode, self._coeffs = state
-        super().__setstate__(superstate)
-
-    @property
-    def coeffs(self, ):
-        return self._coeffs
-
-    @property
-    def detrend_mode(self, ):
-        return self._detrend_mode
-
-    @detrend_mode.setter
-    def detrend_mode(self, detrend_mode):
-        self._detrend_mode = detrend_mode
-        self._detrend()
-
-    @property
-    def is_periodic(self):
-        """
-        SurfaceTopography stays periodic only after detrend mode "center".
-        Otherwise the detrended SurfaceTopography is non-periodic.
-        """
-        if self.detrend_mode == "center":
-            return self.parent_topography.is_periodic
-        else:
-            return False
-
-    @property
-    def x_range(self):
-        return self.parent_topography.x_range
-
-    def positions(self):
-        return self.parent_topography.positions()
-
-    def heights(self):
-        """ Computes the combined profile.
-        """
-        if len(self._coeffs) == 1:
-            a0, = self._coeffs
-            return self.parent_topography.heights() - a0
-        x = self.positions()
-        if len(self._coeffs) == 2:
-            a0, a1 = self._coeffs
-            return self.parent_topography.heights() - a0 - a1 * x
-        elif len(self._coeffs) == 3:
-            a0, a1, a2 = self._coeffs
-            return self.parent_topography.heights() - a0 - a1 * x - a2 * x * x
-        else:
-            raise RuntimeError('Unknown physical_sizes of coefficients '
-                               'tuple.')
-
-    def stringify_plane(self, fmt=lambda x: str(x)):
-        str_coeffs = [fmt(x) for x in self._coeffs]
-        if len(self._coeffs) == 1:
-            h0, = str_coeffs
-            return h0
-        elif len(self._coeffs) == 2:
-            return '{0} + {1} x'.format(*str_coeffs)
-        elif len(self._coeffs) == 3:
-            return '{0} + {1} x + {2} x^2'.format(*str_coeffs)
-        else:
-            raise RuntimeError('Unknown physical_sizes of coefficients '
-                               'tuple.')
-
-    @property
-    def curvatures(self):
-        if len(self._coeffs) == 3:
-            return 2 * self._coeffs[2],
-        else:
-            return 0,
-
-
 # Register analysis functions from this module
 NonuniformLineScanInterface.register_function(
     'mean', lambda this: np.trapz(this.heights(), this.positions()) / this.physical_sizes[0])
@@ -395,4 +261,3 @@ NonuniformLineScanInterface.register_function('max', lambda this: this.heights()
 # Register pipeline functions from this module
 NonuniformLineScanInterface.register_function('to_unit', ScaledNonuniformTopography)
 NonuniformLineScanInterface.register_function('scale', StaticallyScaledNonuniformTopography)
-NonuniformLineScanInterface.register_function('detrend', DetrendedNonuniformTopography)
