+++ test/test_pipeline.py
@@ -29,18 +29,18 @@ Tests of the filter and modification pipeline
 """
 
 import os
-import pytest
 
 import numpy as np
-
+import pytest
 from NuMPI import MPI
 
-from SurfaceTopography.UniformLineScanAndTopography import Topography, \
-    DetrendedUniformTopography, UniformLineScan
 from SurfaceTopography.Generation import fourier_synthesis
 from SurfaceTopography.IO import XYZReader
 from SurfaceTopography.Pipeline import pipeline_function
-from SurfaceTopography.UniformLineScanAndTopography import DecoratedUniformTopography, UniformTopographyInterface
+from SurfaceTopography.Uniform.Detrending import DetrendedUniformTopography
+from SurfaceTopography.UniformLineScanAndTopography import (
+    DecoratedUniformTopography, Topography, UniformLineScan,
+    UniformTopographyInterface)
 
 pytestmark = pytest.mark.skipif(
     MPI.COMM_WORLD.Get_size() > 1,
@@ -119,7 +119,8 @@ def test_uniform_detrended_periodicity():
 
 
 def test_passing_of_docstring():
-    from SurfaceTopography.Uniform.PowerSpectrum import power_spectrum_from_profile
+    from SurfaceTopography.Uniform.PowerSpectrum import \
+        power_spectrum_from_profile
     topography = Topography(np.array([[0, 1, 0], [0, 0, 0]]),
                             physical_sizes=(4., 3.), periodic=True)
     assert topography.power_spectrum_from_profile.__doc__ == power_spectrum_from_profile.__doc__
