+++ test/test_uniform_line_scan.py
@@ -31,7 +31,6 @@ import pickle
 
 import numpy as np
 import pytest
-
 from NuMPI import MPI
 
 from SurfaceTopography import UniformLineScan
@@ -155,7 +154,7 @@ def test_detrend_curvature():
 
     detrended = t.detrend(detrend_mode="curvature")
 
-    assert abs(detrended.coeffs[-1] / detrended.physical_sizes[0] ** 2 - 1 / R) < 1e-12
+    assert abs(detrended.coeffs[-1] - 1 / R) < 1e-12
 
 
 def test_detrend_same_positions():
